chore: move all test helpers into its own package
