+++ internal/server/handlers_accounts_test.go
@@ -1,181 +1,23 @@
 package server_test
 
 import (
-	"encoding/json"
 	"net/http"
-	"path/filepath"
-	"strconv"
-	"strings"
 	"testing"
+
+	tu "github.com/canonical/notary/internal/testutils"
 )
-
-type SuccessResponse struct {
-	Message string `json:"message"`
-}
-
-type GetAccountResponseResult struct {
-	ID          int    `json:"id"`
-	Username    string `json:"username"`
-	Permissions int    `json:"permissions"`
-}
-
-type GetAccountResponse struct {
-	Result GetAccountResponseResult `json:"result"`
-	Error  string                   `json:"error,omitempty"`
-}
-
-type CreateAccountParams struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
-
-type CreateAccountResponseResult struct {
-	ID int `json:"id"`
-}
-
-type CreateAccountResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-type ChangeAccountPasswordParams struct {
-	Password string `json:"password"`
-}
-
-type ChangeAccountPasswordResponseResult struct {
-	ID int `json:"id"`
-}
-
-type ChangeAccountPasswordResponse struct {
-	Result ChangeAccountPasswordResponseResult `json:"result"`
-	Error  string                              `json:"error,omitempty"`
-}
-
-type DeleteAccountResponseResult struct {
-	ID int `json:"id"`
-}
-
-type DeleteAccountResponse struct {
-	Result DeleteAccountResponseResult `json:"result"`
-	Error  string                      `json:"error,omitempty"`
-}
-
-func getAccount(url string, client *http.Client, adminToken string, id int) (int, *GetAccountResponse, error) {
-	req, err := http.NewRequest("GET", url+"/api/v1/accounts/"+strconv.Itoa(id), nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	defer res.Body.Close()
-	var accountResponse GetAccountResponse
-	if err := json.NewDecoder(res.Body).Decode(&accountResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &accountResponse, nil
-}
-
-func getMyAccount(url string, client *http.Client, adminToken string) (int, *GetAccountResponse, error) {
-	req, err := http.NewRequest("GET", url+"/api/v1/accounts/me", nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	defer res.Body.Close()
-	var accountResponse GetAccountResponse
-	if err := json.NewDecoder(res.Body).Decode(&accountResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &accountResponse, nil
-}
-
-func createAccount(url string, client *http.Client, adminToken string, data *CreateAccountParams) (int, *CreateAccountResponse, error) {
-	body, err := json.Marshal(data)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("POST", url+"/api/v1/accounts", strings.NewReader(string(body)))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	defer res.Body.Close()
-	var createResponse CreateAccountResponse
-	if err := json.NewDecoder(res.Body).Decode(&createResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &createResponse, nil
-}
-
-func changeAccountPassword(url string, client *http.Client, adminToken string, id int, data *ChangeAccountPasswordParams) (int, *ChangeAccountPasswordResponse, error) {
-	body, err := json.Marshal(data)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("POST", url+"/api/v1/accounts/"+strconv.Itoa(id)+"/change_password", strings.NewReader(string(body)))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	defer res.Body.Close()
-	var changeResponse ChangeAccountPasswordResponse
-	if err := json.NewDecoder(res.Body).Decode(&changeResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &changeResponse, nil
-}
-
-func deleteAccount(url string, client *http.Client, adminToken string, id int) (int, *DeleteAccountResponse, error) {
-	req, err := http.NewRequest("DELETE", url+"/api/v1/accounts/"+strconv.Itoa(id), nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	defer res.Body.Close()
-	var deleteResponse DeleteAccountResponse
-	if err := json.NewDecoder(res.Body).Decode(&deleteResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &deleteResponse, nil
-}
 
 // This is an end-to-end test for the accounts handlers.
 // The order of the tests is important, as some tests depend on
 // the state of the server after previous tests.
 func TestAccountsEndToEnd(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
 	client := ts.Client()
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
+	nonAdminToken := tu.MustPrepareNonAdminAccount(t, ts, adminToken)
 
 	t.Run("1. Get admin account - admin token", func(t *testing.T) {
-		statusCode, response, err := getAccount(ts.URL, client, adminToken, 1)
+		statusCode, response, err := tu.GetAccount(ts.URL, client, adminToken, 1)
 		if err != nil {
 			t.Fatalf("couldn't get account: %s", err)
 		}
@@ -197,7 +39,7 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("2. Get admin account - non admin token", func(t *testing.T) {
-		statusCode, response, err := getAccount(ts.URL, client, nonAdminToken, 1)
+		statusCode, response, err := tu.GetAccount(ts.URL, client, nonAdminToken, 1)
 		if err != nil {
 			t.Fatalf("couldn't get account: %s", err)
 		}
@@ -210,11 +52,11 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("3. Create account", func(t *testing.T) {
-		createAccountParams := &CreateAccountParams{
+		createAccountParams := &tu.CreateAccountParams{
 			Username: "nopass",
 			Password: "myPassword123!",
 		}
-		statusCode, response, err := createAccount(ts.URL, client, adminToken, createAccountParams)
+		statusCode, response, err := tu.CreateAccount(ts.URL, client, adminToken, createAccountParams)
 		if err != nil {
 			t.Fatalf("couldn't create account: %s", err)
 		}
@@ -227,7 +69,7 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("4. Get account", func(t *testing.T) {
-		statusCode, response, err := getAccount(ts.URL, client, adminToken, 3)
+		statusCode, response, err := tu.GetAccount(ts.URL, client, adminToken, 3)
 		if err != nil {
 			t.Fatalf("couldn't get account: %s", err)
 		}
@@ -249,7 +91,7 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("5. Get account - id not found", func(t *testing.T) {
-		statusCode, response, err := getAccount(ts.URL, client, adminToken, 100)
+		statusCode, response, err := tu.GetAccount(ts.URL, client, adminToken, 100)
 		if err != nil {
 			t.Fatalf("couldn't get account: %s", err)
 		}
@@ -262,10 +104,10 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("6. Change account password - success", func(t *testing.T) {
-		changeAccountPasswordParams := &ChangeAccountPasswordParams{
+		changeAccountPasswordParams := &tu.ChangeAccountPasswordParams{
 			Password: "newPassword1",
 		}
-		statusCode, response, err := changeAccountPassword(ts.URL, client, adminToken, 1, changeAccountPasswordParams)
+		statusCode, response, err := tu.ChangeAccountPassword(ts.URL, client, adminToken, 1, changeAccountPasswordParams)
 		if err != nil {
 			t.Fatalf("couldn't create account: %s", err)
 		}
@@ -278,10 +120,10 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("7. Change account password - no user", func(t *testing.T) {
-		changeAccountPasswordParams := &ChangeAccountPasswordParams{
+		changeAccountPasswordParams := &tu.ChangeAccountPasswordParams{
 			Password: "newPassword1",
 		}
-		statusCode, response, err := changeAccountPassword(ts.URL, client, adminToken, 100, changeAccountPasswordParams)
+		statusCode, response, err := tu.ChangeAccountPassword(ts.URL, client, adminToken, 100, changeAccountPasswordParams)
 		if err != nil {
 			t.Fatalf("couldn't create account: %s", err)
 		}
@@ -294,7 +136,7 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("8. Delete account - success", func(t *testing.T) {
-		statusCode, response, err := deleteAccount(ts.URL, client, adminToken, 2)
+		statusCode, response, err := tu.DeleteAccount(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatalf("couldn't delete account: %s", err)
 		}
@@ -307,7 +149,7 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("9. Delete account - no user", func(t *testing.T) {
-		statusCode, response, err := deleteAccount(ts.URL, client, adminToken, 100)
+		statusCode, response, err := tu.DeleteAccount(ts.URL, client, adminToken, 100)
 		if err != nil {
 			t.Fatalf("couldn't delete account: %s", err)
 		}
@@ -320,7 +162,7 @@ func TestAccountsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("10. Get my admin account - admin token", func(t *testing.T) {
-		statusCode, response, err := getMyAccount(ts.URL, client, adminToken)
+		statusCode, response, err := tu.GetMyAccount(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatalf("couldn't get account: %s", err)
 		}
@@ -343,18 +185,9 @@ func TestAccountsEndToEnd(t *testing.T) {
 }
 
 func TestCreateAccountInvalidInputs(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
 	client := ts.Client()
-
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 
 	tests := []struct {
 		testName string
@@ -384,11 +217,11 @@ func TestCreateAccountInvalidInputs(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.testName, func(t *testing.T) {
-			createAccountParams := &CreateAccountParams{
+			createAccountParams := &tu.CreateAccountParams{
 				Username: test.username,
 				Password: test.password,
 			}
-			statusCode, createCertResponse, err := createAccount(ts.URL, client, adminToken, createAccountParams)
+			statusCode, createCertResponse, err := tu.CreateAccount(ts.URL, client, adminToken, createAccountParams)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -403,18 +236,9 @@ func TestCreateAccountInvalidInputs(t *testing.T) {
 }
 
 func TestChangeAccountPasswordInvalidInputs(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
 	client := ts.Client()
-
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 
 	tests := []struct {
 		testName string
@@ -435,10 +259,10 @@ func TestChangeAccountPasswordInvalidInputs(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.testName, func(t *testing.T) {
-			changeAccountParams := &ChangeAccountPasswordParams{
+			changeAccountParams := &tu.ChangeAccountPasswordParams{
 				Password: test.password,
 			}
-			statusCode, createCertResponse, err := changeAccountPassword(ts.URL, client, adminToken, 1, changeAccountParams)
+			statusCode, createCertResponse, err := tu.ChangeAccountPassword(ts.URL, client, adminToken, 1, changeAccountParams)
 			if err != nil {
 				t.Fatal(err)
 			}
