+++ internal/server/handlers_certificate_authorities_test.go
@@ -1,439 +1,26 @@
 package server_test
 
 import (
-	"bytes"
-	"crypto/rand"
-	"crypto/x509"
-	"encoding/json"
-	"encoding/pem"
 	"fmt"
 	"math/big"
 	"net/http"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/canonical/notary/internal/db"
 	"github.com/canonical/notary/internal/server"
+	tu "github.com/canonical/notary/internal/testutils"
 )
-
-const (
-	selfSignedCACertificatePK = `-----BEGIN RSA PRIVATE KEY-----
-MIIJKAIBAAKCAgEAqzwshq14z01of/TCO+m1XW9k8RLiLce2aVEK1VjuWAR3p8H4
-Oh16TkFUhXnXyY+6gWs+DKKaiX5Jekz1APKGS0gGOMbGWQfjdSP16JHvA80zi6YM
-xV8vc8piCdHzK22F93tTsVzJBL1ApQK+TmhS2WjPKM9tObKafo6uNgHd78eEHVn+
-QBkwvD8HpCnA73LabRyylJ1NC8p+fHT8+t0HjNVZNsEZOzWFZJ755oSZdV949KuY
-VnZDf+Zqmmu5SgwI9ll90hAK2Imk7hEm6t+3ViRdE8MbmLRFBykLhgWZYnJ2wzG9
-/nn4HymJmSt8BzPi+RPfbXOCBvdCRe9MVDtCwHhGukOM3yF/sfQe65EBATBE2PsR
-JWGIoeaaLjK5rT6CvQnc7YKXav8GOQoAn+o8qKuvsf6XItlRALWfwwWmdEfbIX3q
-el+P2hNF4iUoytzQKIBiMUt9IAPwW5Zh7jVK7hm1Sv94P6UIkqbdMgoa/wl9U+vF
-Oi4zgjGU8o3MjyXL716taC/5O8oGBztNDNnc4B4P46+DV/NyGqp5E/DTfZpvslRp
-KUjOmi8g8JaDIypl0m2/825d0+oG5nPrHCQLwXbVWB2UpwfMcF6Jth0SbIM7BG/1
-NnKN5OczIkhoN2ilsRFFNNa+ZEPI4aGzb9J5Hf4d63CTzVtB38Sdnh7rMxUCAwEA
-AQKCAgA3AofTZFtRTa7qnHjhwnzvXV/ySny69FPXlZ+DVqSLRjQigp/6G4o1Jau/
-jZsTN8dU2F0AtiQrU5TLY3m6Ki+Wc7b7+m+yHmSmNz1Cz88XS47pqBimN6QI8NV6
-DiaupurIzKfgbMxvZ1UjLbRxf/ZNvev/UlPmm0girDevRf5Ej7YTr9uMQAt2DlGo
-HXnL8vvU3clJERe1WAI2fWbgOK26Qrf8bSBr8w+9eY0Szzp1iIxVHeM8s5WPuzg4
-D69g6Gjgq3NOTrUNR3riuEPmZKSDWf9E4AM3lYIvgLIhBh92jws0PgCphgl+CVwZ
-Bu07ayHZntbCVL9K+Zgi23OMA9W3dK01f/rTZ4ExfJfN5KUskJ9PfaHK4J/OGAhs
-a0h0CUzhs19uxRPBwjrnIfi2E0EyIqwl+mhDAF3pxLhuTQuVyXYkTER3AQJ1QCGN
-3UP5t/HJKUUxN9jvWvv97XVNVSlkxlWiANKlvu7dvMxiRN9eKqPvzx7R4Oly3Cs8
-jd/X1uw69OFBQ3s+FAtIz8TzKS/ZdbJitZ/sz+L31qrnxb/+E19NQFh0FWElGr22
-pA0ililB4uEJFr4wHGcdyLbdt2vHsZXN5XFkRlFxz2yw57gcffoKz84w2MCtU/Qw
-Z3DgwAFmP8vVpqvTSbiQHSzL2sW/hSN0BCVK+yeULP0dm2d9aQKCAQEAxa2xWcPC
-wKyUI4022kIebd5COLKemmAnW95DShKZJqwyzAqLrHciG5QFvECl3QGmCtoty1nY
-RT1wou6NA6/3yiBJiADqsH/nV6KH1Row9bxt0sVe9ze44tgIxJtqMp4x9E1xWo9L
-zk4jzbufxUz17NvwR9ItiHpDOh88M7U0hBiUxWqBbX/OxCmZTJBdmoGaCk/G/iST
-q7BvNI+4geInE6uzLpfLiWPiFhTSW0D/0Hh9J3hdBDGheuhY7e49RHs5dUsxiVCs
-j6cJ04DvsWyZbiEsAbtO9/eJjGyB+kDUXcrlDDtharhbKxiiXDxEUa+P5+LaANGL
-a++ou4EDsPzI0wKCAQEA3cFBXbtap1dALeM3IMkJQgZ7K7N26tMvbJ8efeJZyu6a
-YsWjOPMjuuY0/tBwlH+Yet6ahIhxQuor0yVRqgEm5uIxl+NYOEw0wBQeCq20YumY
-pNRPb0+zAPQ5UOvoZCiQqwfJzq5czVu0YOx4pW0jbJzJnnhZ0B5p0gIeOgjjVynu
-lkq67c24B5Iy/wiFu/nlGc6/MDtOHpBUvtuWbADIRXptBctICI0+0sHVNMSfeyXf
-czuyPNoWb7Ze1YaB8voPRdT14ObLLGwFCjzirGtu9mFbN4WZemxCvJcm/DWimZob
-WyxXJhDbk++femuCRsS2THkpzvlM/omlEc1RZRcjdwKCAQBDFdQnM4FHZAoOGqFv
-5ppvDMuKdEvQ9irFSaOqYq9o6W1/w2BtUizYER71KTdgzmtsKWj7Ju13agdss+pV
-QwWjqdtqdW0wIuf+3KCeWHofGyhmLCczXMy45znqhxe+P+OSFioO4qyGQgxyiGcL
-TTf0fxuHNDPRqjRgaDNFFQzSe4kZijCMWaBw//EPg7rQcYU2VKainwUicgj7XH7w
-TTCXw2BWwpsHcEdM2Roeb+ug8xL+LyHaB4HWtT1g2cYfFHaGcNNJ14AIbFawWYR+
-wO0867MEj22YR4B9kGF37UJk5jNfOFOyJQiDkqOfC7Dfy+XZeyoitvpK0hWANKNt
-EAyTAoIBAQDN3l360LaGaF/yueAyFbD8lNvAZGnf85Mxej9qirrlxMGbNPQlKMRg
-/NryPTxnAFXkq8gzhh/wCUoKSbkY5Nzit9dmtO7vxP+r2oFRVJYExeyqCVh9dDYw
-ioqzb29dnLNBtIdL01/gXmSFHsOagEnEyYH8Fqr6pWGET+cT5bB4+TrAWDxWDQfu
-L3IchiLxsjtYzDF+a3BIu6GRVna9abSDm/aezGvhVI+gRcbTqD6Oq0hYyHDeQXFW
-4K4F+Uum8TNAl0Z1No4kfVvod0HQ5CQto5B1aRhlKSCDyKeuuXRsuT7wU6fMdRYE
-rw8VLb3SSUkckdeYiKVMISkX787C1MpdAoIBAHw901UFSyy3D78Z0KorBCOi0+ld
-tWr78LXI3aqJk1n3MsybT/9lETx6sE601PW+ObbsSs5y0L1S8vGOH2C4Yjk7Xqdi
-R0kpblMpY6Nh8wyeRn7F9jB4eDOLLYH6793YbjrMzJ14/1fZCfHgy+/xm/QftIP1
-W2OK+pXnlROQiyUPakO5h18x+H1S3+cnQaQXw3S5yRExC6WbmjTC8/q1yB5kEEx7
-cYYP3HUaPT4dKjFM9qt9aF//n8NJGDRw1AVd6MHOPliusM3MDHhRrWsMq2P4yXpM
-9gyV6zS00uKIH4fD5HwldtZjeLGomJK7bkFSMRmMH2jUQvnjtE4pHxJJsUw=
------END RSA PRIVATE KEY-----`
-	selfSignedCACertificate = `-----BEGIN CERTIFICATE-----
-MIIGHzCCBAegAwIBAgIIGCEGnc3aBTkwDQYJKoZIhvcNAQELBQAwgaIxGDAWBgNV
-BAYTD1Rlc3RpbmcgQ291bnRyeTEWMBQGA1UECBMNVGVzdGluZyBTdGF0ZTEZMBcG
-A1UEBxMQVGVzdGluZyBMb2NhbGl0eTEdMBsGA1UEChMUVGVzdGluZyBPcmdhbml6
-YXRpb24xEzARBgNVBAsTClRlc3RpbmcgT1UxHzAdBgNVBAMTFlRlc3RpbmcgU2Vs
-ZiBTaWduZWQgQ0EwHhcNMjUwMjA0MTQxMDA1WhcNMzUwMjA0MTQxMDA1WjCBojEY
-MBYGA1UEBhMPVGVzdGluZyBDb3VudHJ5MRYwFAYDVQQIEw1UZXN0aW5nIFN0YXRl
-MRkwFwYDVQQHExBUZXN0aW5nIExvY2FsaXR5MR0wGwYDVQQKExRUZXN0aW5nIE9y
-Z2FuaXphdGlvbjETMBEGA1UECxMKVGVzdGluZyBPVTEfMB0GA1UEAxMWVGVzdGlu
-ZyBTZWxmIFNpZ25lZCBDQTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIB
-AKs8LIateM9NaH/0wjvptV1vZPES4i3HtmlRCtVY7lgEd6fB+Dodek5BVIV518mP
-uoFrPgyimol+SXpM9QDyhktIBjjGxlkH43Uj9eiR7wPNM4umDMVfL3PKYgnR8ytt
-hfd7U7FcyQS9QKUCvk5oUtlozyjPbTmymn6OrjYB3e/HhB1Z/kAZMLw/B6QpwO9y
-2m0cspSdTQvKfnx0/PrdB4zVWTbBGTs1hWSe+eaEmXVfePSrmFZ2Q3/mappruUoM
-CPZZfdIQCtiJpO4RJurft1YkXRPDG5i0RQcpC4YFmWJydsMxvf55+B8piZkrfAcz
-4vkT321zggb3QkXvTFQ7QsB4RrpDjN8hf7H0HuuRAQEwRNj7ESVhiKHmmi4yua0+
-gr0J3O2Cl2r/BjkKAJ/qPKirr7H+lyLZUQC1n8MFpnRH2yF96npfj9oTReIlKMrc
-0CiAYjFLfSAD8FuWYe41Su4ZtUr/eD+lCJKm3TIKGv8JfVPrxTouM4IxlPKNzI8l
-y+9erWgv+TvKBgc7TQzZ3OAeD+Ovg1fzchqqeRPw032ab7JUaSlIzpovIPCWgyMq
-ZdJtv/NuXdPqBuZz6xwkC8F21VgdlKcHzHBeibYdEmyDOwRv9TZyjeTnMyJIaDdo
-pbERRTTWvmRDyOGhs2/SeR3+Hetwk81bQd/EnZ4e6zMVAgMBAAGjVzBVMA4GA1Ud
-DwEB/wQEAwIBBjATBgNVHSUEDDAKBggrBgEFBQcDATAPBgNVHRMBAf8EBTADAQH/
-MB0GA1UdDgQWBBSbXLqZuUAgHvDfdpdDNDggifOgETANBgkqhkiG9w0BAQsFAAOC
-AgEAKlIcm2sMw6a9PoNFUXLHkGo46Lao1uRJ51tNGBXasbkznb3N8pr0vhVWQk7W
-QM8nd8t8yvaQ7ELo/SapBkjTyu8SwPRVbnXjl1Wke3A34VMusjhIGPNsjgI6zq8i
-rmzzFSNGb+T0AngOc6dMC9N2pAWFHnmzty7Oi39R/jhOmD38wWCMyS1Lui5tXIMJ
-hk/PnVEMF3Rx4rV8FBZh3IJE5O2hB4OCTQVh3w2kOI5+YRWTJm223WfHz7XHR2Ny
-E6yjQhPeglxspeNufhD1H8B49/XuIsHqOBrIPUxztyuBJwXz0KuFitbLrRHLaz2T
-9Ox0i6jEB5f4RU1AiwpM6KE7bFkXmWIYeOykQq7Rj0qGdo50MABGAErIIXpcRzvn
-0r3y1PCvjntdrc9IG8b0bQhVss5Irr9+5wdn90lnKAi2wASmRxbYNVuIi6BpXJvk
-lFukciXd3eW7qZrgUFzS2nBwXv7wlT6e7pfGChSrE+dlh6QBHkDlHU2FdwdAc8BX
-GbaLCYgW2R7/cgFG25APKkLxVzuIVYCQzIc94vyTqdhEhdQfvDTnnE/Nj//lVJBp
-vpdEAlIxCc8EPphBf2okYk9nbda/3fTIBlYhXo3jGw/luYY+rv3YBMzPJ9KXkfxH
-nCSo5Rj3yTrtNYFLnu+iLCvMb5PcJXE55Pu5OYGktHnvMgc=
------END CERTIFICATE-----`
-)
-
-type CreateCertificateAuthorityParams struct {
-	SelfSigned bool `json:"self_signed"`
-
-	CommonName          string `json:"common_name"`
-	SANsDNS             string `json:"sans_dns"`
-	CountryName         string `json:"country_name"`
-	StateOrProvinceName string `json:"state_or_province_name"`
-	LocalityName        string `json:"locality_name"`
-	OrganizationName    string `json:"organization_name"`
-	OrganizationalUnit  string `json:"organizational_unit_name"`
-	NotValidAfter       string `json:"not_valid_after"`
-}
-
-type CreateCertificateAuthorityResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func createCertificateAuthority(url string, client *http.Client, adminToken string, ca CreateCertificateAuthorityParams) (int, *CreateCertificateAuthorityResponse, error) {
-	reqData, err := json.Marshal(ca)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("POST", url+"/api/v1/certificate_authorities", bytes.NewReader(reqData))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	req.Header.Set("Content-Type", "application/json")
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var createCertificateAuthorityResponse CreateCertificateAuthorityResponse
-	if err := json.NewDecoder(res.Body).Decode(&createCertificateAuthorityResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &createCertificateAuthorityResponse, nil
-}
-
-type ListCertificateAuthoritiesResponse struct {
-	Result []server.CertificateAuthority `json:"result"`
-	Error  string                        `json:"error,omitempty"`
-}
-
-func listCertificateAuthorities(url string, client *http.Client, adminToken string) (int, *ListCertificateAuthoritiesResponse, error) {
-	req, err := http.NewRequest("GET", url+"/api/v1/certificate_authorities", nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var certificateAuthoritiesResponse ListCertificateAuthoritiesResponse
-	if err := json.NewDecoder(res.Body).Decode(&certificateAuthoritiesResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &certificateAuthoritiesResponse, nil
-}
-
-type GetCertificateAuthorityResponse struct {
-	Result server.CertificateAuthority `json:"result"`
-	Error  string                      `json:"error,omitempty"`
-}
-
-func getCertificateAuthority(url string, client *http.Client, adminToken string, id int) (int, *GetCertificateAuthorityResponse, error) {
-	req, err := http.NewRequest("GET", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id), nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var getCertificateAuthorityResponse GetCertificateAuthorityResponse
-	if err := json.NewDecoder(res.Body).Decode(&getCertificateAuthorityResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &getCertificateAuthorityResponse, nil
-}
-
-type UpdateCertificateAuthorityParams struct {
-	Status string `json:"status,omitempty"`
-}
-
-type UpdateCertificateAuthorityResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func updateCertificateAuthority(url string, client *http.Client, adminToken string, id int, status UpdateCertificateAuthorityParams) (int, *UpdateCertificateAuthorityResponse, error) {
-	reqData, err := json.Marshal(status)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("PUT", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id), bytes.NewReader(reqData))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	req.Header.Set("Content-Type", "application/json")
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var updateCertificateAuthorityResponse UpdateCertificateAuthorityResponse
-	if err := json.NewDecoder(res.Body).Decode(&updateCertificateAuthorityResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &updateCertificateAuthorityResponse, nil
-}
-
-func deleteCertificateAuthority(url string, client *http.Client, adminToken string, id int) (int, error) {
-	req, err := http.NewRequest("DELETE", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id), nil)
-	if err != nil {
-		return 0, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, err
-	}
-	return res.StatusCode, nil
-}
-
-type UploadCertificateToCertificateAuthorityResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func uploadCertificateToCertificateAuthority(url string, client *http.Client, adminToken string, id int, cert server.UploadCertificateToCertificateAuthorityParams) (int, *UploadCertificateToCertificateAuthorityResponse, error) {
-	reqData, err := json.Marshal(cert)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("POST", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id)+"/certificate", bytes.NewReader(reqData))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	req.Header.Set("Content-Type", "application/json")
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var uploadCertificateToCertificateAuthorityResponse UploadCertificateToCertificateAuthorityResponse
-	if err := json.NewDecoder(res.Body).Decode(&uploadCertificateToCertificateAuthorityResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &uploadCertificateToCertificateAuthorityResponse, nil
-}
-
-type SignCertificateRequestResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func signCertificateRequest(url string, client *http.Client, adminToken string, id int, cert server.SignCertificateRequestParams) (int, *SignCertificateRequestResponse, error) {
-	reqData, err := json.Marshal(cert)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("POST", url+"/api/v1/certificate_requests/"+strconv.Itoa(id)+"/sign", bytes.NewReader(reqData))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	req.Header.Set("Content-Type", "application/json")
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var signCertificateRequestResponse SignCertificateRequestResponse
-	if err := json.NewDecoder(res.Body).Decode(&signCertificateRequestResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &signCertificateRequestResponse, nil
-}
-
-type SignCertificateAuthorityResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func signCertificateAuthority(url string, client *http.Client, adminToken string, id int, cert server.SignCertificateAuthorityParams) (int, *SignCertificateAuthorityResponse, error) {
-	reqData, err := json.Marshal(cert)
-	if err != nil {
-		return 0, nil, err
-	}
-	req, err := http.NewRequest("POST", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id)+"/sign", bytes.NewReader(reqData))
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	req.Header.Set("Content-Type", "application/json")
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var signCertificateAuthorityResponse SignCertificateAuthorityResponse
-	if err := json.NewDecoder(res.Body).Decode(&signCertificateAuthorityResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &signCertificateAuthorityResponse, nil
-}
-
-type RevokeCertificateAuthorityCertificateResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func revokeCertificateAuthority(url string, client *http.Client, adminToken string, id int) (int, *RevokeCertificateAuthorityCertificateResponse, error) {
-	req, err := http.NewRequest("POST", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id)+"/revoke", nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var RevokeCertificateAuthorityResponse RevokeCertificateAuthorityCertificateResponse
-	if err := json.NewDecoder(res.Body).Decode(&RevokeCertificateAuthorityResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &RevokeCertificateAuthorityResponse, nil
-}
-
-type RevokeCertificateRequestResponse struct {
-	Result SuccessResponse `json:"result"`
-	Error  string          `json:"error,omitempty"`
-}
-
-func revokeCertificateRequest(url string, client *http.Client, adminToken string, id int) (int, *RevokeCertificateRequestResponse, error) {
-	req, err := http.NewRequest("POST", url+"/api/v1/certificate_requests/"+strconv.Itoa(id)+"/certificate/revoke", nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var RevokeCertificateRequestResponse RevokeCertificateRequestResponse
-	if err := json.NewDecoder(res.Body).Decode(&RevokeCertificateRequestResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &RevokeCertificateRequestResponse, nil
-}
-
-type GetCRLResponse struct {
-	Result server.CRL `json:"result"`
-	Error  string     `json:"error,omitempty"`
-}
-
-func getCertificateAuthorityCRLRequest(url string, client *http.Client, adminToken string, id int) (int, *GetCRLResponse, error) {
-	req, err := http.NewRequest("GET", url+"/api/v1/certificate_authorities/"+strconv.Itoa(id)+"/crl", nil)
-	if err != nil {
-		return 0, nil, err
-	}
-	req.Header.Set("Authorization", "Bearer "+adminToken)
-	res, err := client.Do(req)
-	if err != nil {
-		return 0, nil, err
-	}
-	var GetCRLResponse GetCRLResponse
-	if err := json.NewDecoder(res.Body).Decode(&GetCRLResponse); err != nil {
-		return 0, nil, err
-	}
-	return res.StatusCode, &GetCRLResponse, nil
-}
-
-// sign a csr with a self signed ca
-func signCSR(csr string) string {
-	csrDER, _ := pem.Decode([]byte(csr))                             //nolint: errcheck
-	csrTemplate, _ := x509.ParseCertificateRequest(csrDER.Bytes)     //nolint: errcheck
-	signingCertDER, _ := pem.Decode([]byte(selfSignedCACertificate)) //nolint: errcheck
-	signingCert, _ := x509.ParseCertificate(signingCertDER.Bytes)    //nolint: errcheck
-	pkDER, _ := pem.Decode([]byte(selfSignedCACertificatePK))        //nolint: errcheck
-	pk, _ := x509.ParsePKCS1PrivateKey(pkDER.Bytes)                  //nolint: errcheck
-
-	certTemplate := x509.Certificate{
-		SerialNumber:          big.NewInt(time.Now().UnixNano()),
-		Subject:               csrTemplate.Subject,
-		DNSNames:              csrTemplate.DNSNames,
-		EmailAddresses:        csrTemplate.EmailAddresses,
-		IPAddresses:           csrTemplate.IPAddresses,
-		URIs:                  csrTemplate.URIs,
-		NotBefore:             time.Now(),
-		NotAfter:              time.Now().AddDate(1, 0, 0),
-		BasicConstraintsValid: true,
-		IsCA:                  true,
-		KeyUsage:              x509.KeyUsageCertSign | x509.KeyUsageCRLSign,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
-	}
-
-	finalCert, _ := x509.CreateCertificate(rand.Reader, &certTemplate, signingCert, csrTemplate.PublicKey, pk) //nolint: errcheck
-	certPEM := new(bytes.Buffer)
-	pem.Encode(certPEM, &pem.Block{ //nolint: errcheck
-		Type:  "CERTIFICATE",
-		Bytes: finalCert,
-	})
-
-	return certPEM.String()
-}
 
 // The order of the tests is important, as some tests depend on the state of the server after previous tests.
 func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 	client := ts.Client()
 
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
 	t.Run("1. List certificate authorities", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -449,7 +36,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 	})
 
 	t.Run("2. Create self signed certificate authority", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: true,
 
 			CommonName:          "Self Signed CA",
@@ -461,7 +48,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -474,7 +61,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 	})
 
 	t.Run("3. Get all CA's - 1 should be there and enabled", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -496,7 +83,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 	})
 
 	t.Run("4. Make a new Intermediate CA", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: false,
 
 			CommonName:          "Not Self Signed CA",
@@ -508,7 +95,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -520,7 +107,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("5. Get all CA's - 2 should be there, one enabled one pending", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -543,7 +130,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 
 	var IntermediateCACSR string
 	t.Run("6. Get CA by ID", func(t *testing.T) {
-		statusCode, getCAResponse, err := getCertificateAuthority(ts.URL, client, adminToken, 2)
+		statusCode, getCAResponse, err := tu.GetCertificateAuthority(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -569,8 +156,8 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 	})
 
 	t.Run("7. Sign the intermediate CA's CSR", func(t *testing.T) {
-		signedCert := signCSR(IntermediateCACSR)
-		statusCode, uploadCertificateResponse, err := uploadCertificateToCertificateAuthority(ts.URL, client, adminToken, 2, server.UploadCertificateToCertificateAuthorityParams{CertificateChain: signedCert + selfSignedCACertificate})
+		signedCert := tu.SignCSR(IntermediateCACSR)
+		statusCode, uploadCertificateResponse, err := tu.UploadCertificateToCertificateAuthority(ts.URL, client, adminToken, 2, server.UploadCertificateToCertificateAuthorityParams{CertificateChain: signedCert + tu.SelfSignedCACertificate})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -582,7 +169,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("8. Get all CA's - 2 should be there and both enabled", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -603,7 +190,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("9. Make first CA legacy", func(t *testing.T) {
-		statusCode, makeLegacyResponse, err := updateCertificateAuthority(ts.URL, client, adminToken, 1, UpdateCertificateAuthorityParams{Status: "legacy"})
+		statusCode, makeLegacyResponse, err := tu.UpdateCertificateAuthority(ts.URL, client, adminToken, 1, tu.UpdateCertificateAuthorityParams{Status: "legacy"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -615,7 +202,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("10. Get all CA's - 1 enabled 1 disabled should be there", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -636,7 +223,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("11. Delete first CA", func(t *testing.T) {
-		statusCode, err := deleteCertificateAuthority(ts.URL, client, adminToken, 1)
+		statusCode, err := tu.DeleteCertificateAuthority(ts.URL, client, adminToken, 1)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -645,7 +232,7 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("12. Get all CA's - 1 enabled should be there", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -665,18 +252,9 @@ func TestSelfSignedCertificateAuthorityEndToEnd(t *testing.T) {
 }
 
 func TestCreateCertificateAuthorityInvalidInputs(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 	client := ts.Client()
-
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
 
 	tests := []struct {
 		testName            string
@@ -737,7 +315,7 @@ func TestCreateCertificateAuthorityInvalidInputs(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.testName, func(t *testing.T) {
-			createCertificateAuthorityRequest := CreateCertificateAuthorityParams{
+			createCertificateAuthorityRequest := tu.CreateCertificateAuthorityParams{
 				SelfSigned:          test.selfSigned,
 				CommonName:          test.commonName,
 				SANsDNS:             test.sansDNS,
@@ -748,7 +326,7 @@ func TestCreateCertificateAuthorityInvalidInputs(t *testing.T) {
 				OrganizationalUnit:  test.organizationalUnit,
 				NotValidAfter:       test.notValidAfter,
 			}
-			statusCode, createCertResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificateAuthorityRequest)
+			statusCode, createCertResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificateAuthorityRequest)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -763,19 +341,11 @@ func TestCreateCertificateAuthorityInvalidInputs(t *testing.T) {
 }
 
 func TestUploadCertificateToCertificateAuthorityInvalidInputs(t *testing.T) {
-	tempDir := t.TempDir()
-	dbPath := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(dbPath)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 	client := ts.Client()
 
-	var adminToken, nonAdminToken string
-	t.Run("prepare accounts", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
-	createCAParams := CreateCertificateAuthorityParams{
+	createCAParams := tu.CreateCertificateAuthorityParams{
 		SelfSigned:          false,
 		CommonName:          "Intermediate CA",
 		SANsDNS:             "intermediate.example.com",
@@ -786,7 +356,7 @@ func TestUploadCertificateToCertificateAuthorityInvalidInputs(t *testing.T) {
 		OrganizationalUnit:  "Testing",
 		NotValidAfter:       time.Now().AddDate(5, 0, 0).Format(time.RFC3339),
 	}
-	statusCode, _, err := createCertificateAuthority(ts.URL, client, adminToken, createCAParams)
+	statusCode, _, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCAParams)
 	if err != nil {
 		t.Fatalf("error creating certificate authority: %v", err)
 	}
@@ -830,7 +400,7 @@ invalid
 			uploadParams := server.UploadCertificateToCertificateAuthorityParams{
 				CertificateChain: tc.certificateChain,
 			}
-			statusCode, uploadResponse, err := uploadCertificateToCertificateAuthority(ts.URL, client, adminToken, 1, uploadParams)
+			statusCode, uploadResponse, err := tu.UploadCertificateToCertificateAuthority(ts.URL, client, adminToken, 1, uploadParams)
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
 			}
@@ -845,21 +415,12 @@ invalid
 }
 
 func TestSignCertificatesEndToEnd(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 	client := ts.Client()
 
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
 	t.Run("1. List certificate authorities", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -875,7 +436,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 	})
 
 	t.Run("2. Create self signed certificate authority", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: true,
 
 			CommonName:          "Self Signed CA",
@@ -887,7 +448,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -900,7 +461,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 	})
 
 	t.Run("3. Get all CA's - 1 should be there and enabled", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -922,7 +483,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 	})
 
 	t.Run("4. Make a new Intermediate CA", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: false,
 
 			CommonName:          "Intermediate CA",
@@ -933,7 +494,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -945,7 +506,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("5. Get all CA's - 2 should be there, one enabled one disabled", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -967,7 +528,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 	})
 
 	t.Run("6. Get CA by ID", func(t *testing.T) {
-		statusCode, getCAResponse, err := getCertificateAuthority(ts.URL, client, adminToken, 2)
+		statusCode, getCAResponse, err := tu.GetCertificateAuthority(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -991,8 +552,8 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("7. Try signing CSR with unsigned intermediate CA - should fail", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: csr1}
-		statusCode, createCertResponse, err := createCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: tu.AppleCSR}
+		statusCode, createCertResponse, err := tu.CreateCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1002,7 +563,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		if createCertResponse.Error != "" {
 			t.Fatalf("expected no error, got %s", createCertResponse.Error)
 		}
-		statusCode, signCertificateRequestResponse, err := signCertificateRequest(ts.URL, client, adminToken, 3, server.SignCertificateRequestParams{CertificateAuthorityID: "2"})
+		statusCode, signCertificateRequestResponse, err := tu.SignCertificateRequest(ts.URL, client, adminToken, 3, server.SignCertificateRequestParams{CertificateAuthorityID: "2"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1015,7 +576,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 	})
 
 	t.Run("8. Sign the intermediate CA's CSR", func(t *testing.T) {
-		statusCode, uploadCertificateResponse, err := signCertificateAuthority(ts.URL, client, adminToken, 2, server.SignCertificateAuthorityParams{CertificateAuthorityID: "1"})
+		statusCode, uploadCertificateResponse, err := tu.SignCertificateAuthority(ts.URL, client, adminToken, 2, server.SignCertificateAuthorityParams{CertificateAuthorityID: "1"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1027,7 +588,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("9. Get all CA's - 2 should be there and both active", func(t *testing.T) {
-		statusCode, listCAsResponse, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, listCAsResponse, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1051,8 +612,8 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("10. Create 2nd CSR's", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: csr2}
-		statusCode, createCertResponse, err := createCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: tu.StrawberryCSR}
+		statusCode, createCertResponse, err := tu.CreateCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1064,7 +625,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("11. Try Signing a CA CSR - should fail", func(t *testing.T) {
-		statusCode, signCertificateRequestResponse, err := signCertificateRequest(ts.URL, client, adminToken, 1, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
+		statusCode, signCertificateRequestResponse, err := tu.SignCertificateRequest(ts.URL, client, adminToken, 1, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1076,7 +637,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("12. Sign CSRs with each CA", func(t *testing.T) {
-		statusCode, signCertificateRequestResponse, err := signCertificateRequest(ts.URL, client, adminToken, 3, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
+		statusCode, signCertificateRequestResponse, err := tu.SignCertificateRequest(ts.URL, client, adminToken, 3, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1086,7 +647,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		if signCertificateRequestResponse.Error != "" {
 			t.Fatalf("expected success, got %s", signCertificateRequestResponse.Error)
 		}
-		statusCode, signCertificateRequestResponse, err = signCertificateRequest(ts.URL, client, adminToken, 4, server.SignCertificateRequestParams{CertificateAuthorityID: "2"})
+		statusCode, signCertificateRequestResponse, err = tu.SignCertificateRequest(ts.URL, client, adminToken, 4, server.SignCertificateRequestParams{CertificateAuthorityID: "2"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1098,7 +659,7 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("13. Validate CSRs", func(t *testing.T) {
-		statusCode, listCSRsResponse, err := listCertificateRequests(ts.URL, client, adminToken)
+		statusCode, listCSRsResponse, err := tu.ListCertificateRequests(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1127,21 +688,12 @@ func TestSignCertificatesEndToEnd(t *testing.T) {
 }
 
 func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 	client := ts.Client()
 
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
 	t.Run("1. Create self signed certificate authority", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: true,
 
 			CommonName:          "Self Signed CA",
@@ -1153,7 +705,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1166,7 +718,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 	})
 
 	t.Run("2. Create Intermediate certificate authority", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: false,
 
 			CommonName:          "Not Self Signed CA",
@@ -1178,7 +730,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1191,8 +743,8 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 	})
 
 	t.Run("3. Create CSR", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: csr1}
-		statusCode, createCSRResponse, err := createCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: tu.AppleCSR}
+		statusCode, createCSRResponse, err := tu.CreateCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1205,7 +757,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 	})
 
 	t.Run("4. Get CSRs - only 1 should appear", func(t *testing.T) {
-		statusCode, listCertsResponse, err := listCertificateRequests(ts.URL, client, adminToken)
+		statusCode, listCertsResponse, err := tu.ListCertificateRequests(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1223,7 +775,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 		}
 	})
 	t.Run("5. Get CSR - should fail", func(t *testing.T) {
-		statusCode, getCertResponse, err := getCertificateRequest(ts.URL, client, adminToken, 2)
+		statusCode, getCertResponse, err := tu.GetCertificateRequest(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1235,7 +787,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 		}
 	})
 	t.Run("6. Delete CA CSR - should fail", func(t *testing.T) {
-		statusCode, err := deleteCertificateRequest(ts.URL, client, adminToken, 2)
+		statusCode, err := tu.DeleteCertificateRequest(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1244,7 +796,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 		}
 	})
 	t.Run("7. Reject CA CSR - should fail", func(t *testing.T) {
-		statusCode, err := rejectCertificate(ts.URL, client, adminToken, 2)
+		statusCode, err := tu.RejectCertificate(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1253,7 +805,7 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 		}
 	})
 	t.Run("8. Sign CA CSR - should fail", func(t *testing.T) {
-		statusCode, signCertificateRequestResponse, err := signCertificateRequest(ts.URL, client, adminToken, 1, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
+		statusCode, signCertificateRequestResponse, err := tu.SignCertificateRequest(ts.URL, client, adminToken, 1, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1267,21 +819,12 @@ func TestUnsuccessfulRequestsMadeToCACSRs(t *testing.T) {
 }
 
 func TestCertificateRevocationListsEndToEnd(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
+	ts := tu.MustPrepareServer(t)
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
 	client := ts.Client()
 
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare user accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
 	t.Run("1. Create self signed certificate authority", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: true,
 
 			CommonName:          "Self Signed CA",
@@ -1292,7 +835,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1304,7 +847,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("2. Create Intermediate CA", func(t *testing.T) {
-		createCertificatAuthorityParams := CreateCertificateAuthorityParams{
+		createCertificatAuthorityParams := tu.CreateCertificateAuthorityParams{
 			SelfSigned: false,
 
 			CommonName:          "Intermediate CA",
@@ -1315,7 +858,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 			OrganizationalUnit:  "Identity",
 			NotValidAfter:       "2030-01-01T00:00:00Z",
 		}
-		statusCode, createCAResponse, err := createCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
+		statusCode, createCAResponse, err := tu.CreateCertificateAuthority(ts.URL, client, adminToken, createCertificatAuthorityParams)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1328,7 +871,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("3. Get CA CRL's. Root CA should have one, Intermediate shouldn't", func(t *testing.T) {
-		statusCode, cas, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, cas, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if statusCode != http.StatusOK {
 			t.Fatalf("expected status %d, got %d", http.StatusOK, statusCode)
 		}
@@ -1347,7 +890,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("4. Sign Intermediate CA", func(t *testing.T) {
-		statusCode, signCAResponse, err := signCertificateAuthority(ts.URL, client, adminToken, 2, server.SignCertificateAuthorityParams{CertificateAuthorityID: "1"})
+		statusCode, signCAResponse, err := tu.SignCertificateAuthority(ts.URL, client, adminToken, 2, server.SignCertificateAuthorityParams{CertificateAuthorityID: "1"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1360,7 +903,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("5. Get CA CRL's. Both should have one.", func(t *testing.T) {
-		statusCode, cas, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, cas, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if statusCode != http.StatusOK {
 			t.Fatalf("expected status %d, got %d", http.StatusOK, statusCode)
 		}
@@ -1380,8 +923,8 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 
 	t.Run("6. Add 2 CSR's and sign them.", func(t *testing.T) {
 
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: csr1}
-		statusCode, createCertResponse, err := createCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: tu.AppleCSR}
+		statusCode, createCertResponse, err := tu.CreateCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1391,8 +934,8 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		if createCertResponse.Error != "" {
 			t.Fatalf("expected no error, got %s", createCertResponse.Error)
 		}
-		createCertificateRequestRequest = CreateCertificateRequestParams{CSR: csr2}
-		statusCode, createCertResponse, err = createCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
+		createCertificateRequestRequest = tu.CreateCertificateRequestParams{CSR: tu.StrawberryCSR}
+		statusCode, createCertResponse, err = tu.CreateCertificateRequest(ts.URL, client, adminToken, createCertificateRequestRequest)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -1402,7 +945,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		if createCertResponse.Error != "" {
 			t.Fatalf("expected no error, got %s", createCertResponse.Error)
 		}
-		statusCode, signCertificateRequestResponse, err := signCertificateRequest(ts.URL, client, adminToken, 3, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
+		statusCode, signCertificateRequestResponse, err := tu.SignCertificateRequest(ts.URL, client, adminToken, 3, server.SignCertificateRequestParams{CertificateAuthorityID: "1"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1412,7 +955,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		if signCertificateRequestResponse.Error != "" {
 			t.Fatalf("expected success, got %s", signCertificateRequestResponse.Error)
 		}
-		statusCode, signCertificateRequestResponse, err = signCertificateRequest(ts.URL, client, adminToken, 4, server.SignCertificateRequestParams{CertificateAuthorityID: "2"})
+		statusCode, signCertificateRequestResponse, err = tu.SignCertificateRequest(ts.URL, client, adminToken, 4, server.SignCertificateRequestParams{CertificateAuthorityID: "2"})
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1424,7 +967,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("7. Get CSR's. Both should have the correct CRLDistributionPoint extension.", func(t *testing.T) {
-		statusCode, listCSRsResponse, err := listCertificateRequests(ts.URL, client, adminToken)
+		statusCode, listCSRsResponse, err := tu.ListCertificateRequests(ts.URL, client, adminToken)
 		if statusCode != http.StatusOK {
 			t.Fatalf("expected status %d, got %d", http.StatusOK, statusCode)
 		}
@@ -1452,7 +995,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("8. Revoke both certificates.", func(t *testing.T) {
-		statusCode, response, err := revokeCertificateRequest(ts.URL, client, adminToken, 3)
+		statusCode, response, err := tu.RevokeCertificateRequest(ts.URL, client, adminToken, 3)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1462,7 +1005,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		if response.Error != "" {
 			t.Fatalf("expected success, got %s", response.Error)
 		}
-		statusCode, response, err = revokeCertificateRequest(ts.URL, client, adminToken, 4)
+		statusCode, response, err = tu.RevokeCertificateRequest(ts.URL, client, adminToken, 4)
 		if err != nil {
 			t.Fatal("expected no error, got: ", err)
 		}
@@ -1472,7 +1015,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		if response.Error != "" {
 			t.Fatalf("expected success, got %s", response.Error)
 		}
-		statusCode, listCSRsResponse, err := listCertificateRequests(ts.URL, client, adminToken)
+		statusCode, listCSRsResponse, err := tu.ListCertificateRequests(ts.URL, client, adminToken)
 		if statusCode != http.StatusOK {
 			t.Fatalf("expected status %d, got %d", http.StatusOK, statusCode)
 		}
@@ -1487,7 +1030,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 				t.Fatalf("expected no certificate, got '%s'", csr.CertificateChain)
 			}
 		}
-		statusCode, cas, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, cas, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if err != nil {
 			t.Fatalf("expected no error checking CA status, got: %s", err)
 		}
@@ -1506,7 +1049,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("9. Get both CA's. Each CA should have 1 certificate in their CRL", func(t *testing.T) {
-		statusCode, cas, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, cas, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if statusCode != http.StatusOK {
 			t.Fatalf("expected status %d, got %d", http.StatusOK, statusCode)
 		}
@@ -1531,7 +1074,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 	})
 
 	t.Run("10. Revoke Intermediate CA", func(t *testing.T) {
-		statusCode, response, err := revokeCertificateAuthority(ts.URL, client, adminToken, 2)
+		statusCode, response, err := tu.RevokeCertificateAuthority(ts.URL, client, adminToken, 2)
 		if err != nil {
 			t.Fatalf("expected no error, got: %s", err)
 		}
@@ -1541,7 +1084,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		if response.Error != "" {
 			t.Fatalf("expected success, got %s", response.Error)
 		}
-		statusCode, cas, err := listCertificateAuthorities(ts.URL, client, adminToken)
+		statusCode, cas, err := tu.ListCertificateAuthorities(ts.URL, client, adminToken)
 		if statusCode != http.StatusOK {
 			t.Fatalf("expected status %d, got %d", http.StatusOK, statusCode)
 		}
@@ -1563,7 +1106,7 @@ func TestCertificateRevocationListsEndToEnd(t *testing.T) {
 		}
 	})
 	t.Run("11. Get CRL as a non-authenticated user", func(t *testing.T) {
-		statusCode, result, err := getCertificateAuthorityCRLRequest(ts.URL, client, "", 1)
+		statusCode, result, err := tu.GetCertificateAuthorityCRLRequest(ts.URL, client, "", 1)
 		if err != nil {
 			t.Fatalf("expected no error, got: %s", err)
 		}
