+++ internal/server/server_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/canonical/notary/internal/encryption_backend"
 	"github.com/canonical/notary/internal/server"
+	tu "github.com/canonical/notary/internal/testutils"
 	"go.uber.org/zap"
 )
 
@@ -21,7 +22,7 @@ func TestNewSuccess(t *testing.T) {
 		t.Fatalf("cannot create logger: %s", err)
 	}
 	noneEncryptionBackend := encryption_backend.NoEncryptionBackend{}
-  s, err := server.New(8000, []byte(cert), []byte(key), db_path, "example.com", false, l, noneEncryptionBackend)
+	s, err := server.New(8000, []byte(tu.TestServerCertificate), []byte(tu.TestServerKey), db_path, "example.com", false, l, noneEncryptionBackend)
 	if err != nil {
 		t.Errorf("Error occurred: %s", err)
 	}
@@ -36,13 +37,64 @@ func TestInvalidKeyFailure(t *testing.T) {
 		t.Fatalf("cannot create logger: %s", err)
 	}
 	noneEncryptionBackend := encryption_backend.NoEncryptionBackend{}
-	_, err = server.New(8000, []byte(cert), []byte{}, "notary.db", "example.com", false, l, noneEncryptionBackend)
+	_, err = server.New(8000, []byte(tu.TestServerCertificate), []byte{}, "notary.db", "example.com", false, l, noneEncryptionBackend)
 	if err == nil {
 		t.Errorf("No error was thrown for invalid key")
 	}
 }
 
-func createRequestBomb(url string, client *http.Client, adminToken string, certRequest CreateCertificateRequestParams) (int, error) {
+func TestRequestOverload(t *testing.T) {
+	ts := tu.MustPrepareServer(t)
+	client := ts.Client()
+	adminToken := tu.MustPrepareAdminAccount(t, ts)
+
+	t.Run("throw a valid size string", func(t *testing.T) {
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: generateRandomString(20)}
+		statusCode, err := createRequestBomb(ts.URL, client, adminToken, createCertificateRequestRequest)
+		if err != nil {
+			t.Fatalf("couldn't get status: %s", err)
+		}
+
+		if statusCode != http.StatusBadRequest {
+			t.Fatalf("expected status %d, got %d", http.StatusBadRequest, statusCode)
+		}
+	})
+
+	t.Run("throw a bomb", func(t *testing.T) {
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: generateRandomString(200)}
+		statusCode, err := createRequestBomb(ts.URL, client, adminToken, createCertificateRequestRequest)
+		if err != nil {
+			t.Fatalf("couldn't get status: %s", err)
+		}
+		if statusCode != http.StatusRequestEntityTooLarge {
+			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, statusCode)
+		}
+	})
+
+	t.Run("throw a bomb with no content length header", func(t *testing.T) {
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: generateRandomString(200)}
+		statusCode, err := createRequestBombWithCustomHeader(ts.URL, client, adminToken, createCertificateRequestRequest, "")
+		if err != nil {
+			t.Fatalf("couldn't get status: %s", err)
+		}
+		if statusCode != http.StatusRequestEntityTooLarge {
+			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, statusCode)
+		}
+	})
+
+	t.Run("throw a bomb with fake content length header", func(t *testing.T) {
+		createCertificateRequestRequest := tu.CreateCertificateRequestParams{CSR: generateRandomString(200)}
+		statusCode, err := createRequestBombWithCustomHeader(ts.URL, client, adminToken, createCertificateRequestRequest, "2")
+		if err != nil {
+			t.Fatalf("couldn't get status: %s", err)
+		}
+		if statusCode != http.StatusRequestEntityTooLarge {
+			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, statusCode)
+		}
+	})
+}
+
+func createRequestBomb(url string, client *http.Client, adminToken string, certRequest tu.CreateCertificateRequestParams) (int, error) {
 	reqData, err := json.Marshal(certRequest)
 	if err != nil {
 		return 0, err
@@ -57,14 +109,14 @@ func createRequestBomb(url string, client *http.Client, adminToken string, certR
 	if err != nil {
 		return 0, err
 	}
-	var createCertificateRequestResponse CreateCertificateRequestResponse
+	var createCertificateRequestResponse tu.CreateCertificateRequestResponse
 	if err := json.NewDecoder(res.Body).Decode(&createCertificateRequestResponse); err != nil {
 		return 0, err
 	}
 	return res.StatusCode, nil
 }
 
-func createRequestBombWithCustomHeader(url string, client *http.Client, adminToken string, certRequest CreateCertificateRequestParams, contentLengthHeaderData string) (int, error) {
+func createRequestBombWithCustomHeader(url string, client *http.Client, adminToken string, certRequest tu.CreateCertificateRequestParams, contentLengthHeaderData string) (int, error) {
 	reqData, err := json.Marshal(certRequest)
 	if err != nil {
 		return 0, err
@@ -80,71 +132,11 @@ func createRequestBombWithCustomHeader(url string, client *http.Client, adminTok
 	if err != nil {
 		return 0, err
 	}
-	var createCertificateRequestResponse CreateCertificateRequestResponse
+	var createCertificateRequestResponse tu.CreateCertificateRequestResponse
 	if err := json.NewDecoder(res.Body).Decode(&createCertificateRequestResponse); err != nil {
 		return 0, err
 	}
 	return res.StatusCode, nil
-}
-
-func TestRequestOverload(t *testing.T) {
-	tempDir := t.TempDir()
-	db_path := filepath.Join(tempDir, "db.sqlite3")
-	ts, _, err := setupServer(db_path)
-	if err != nil {
-		t.Fatalf("couldn't create test server: %s", err)
-	}
-	defer ts.Close()
-	client := ts.Client()
-
-	var adminToken string
-	var nonAdminToken string
-	t.Run("prepare accounts and tokens", prepareAccounts(ts.URL, client, &adminToken, &nonAdminToken))
-
-	t.Run("throw a valid size string", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: generateRandomString(20)}
-		statusCode, err := createRequestBomb(ts.URL, client, adminToken, createCertificateRequestRequest)
-		if err != nil {
-			t.Fatalf("couldn't get status: %s", err)
-		}
-
-		if statusCode != http.StatusBadRequest {
-			t.Fatalf("expected status %d, got %d", http.StatusBadRequest, statusCode)
-		}
-	})
-
-	t.Run("throw a bomb", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: generateRandomString(200)}
-		statusCode, err := createRequestBomb(ts.URL, client, adminToken, createCertificateRequestRequest)
-		if err != nil {
-			t.Fatalf("couldn't get status: %s", err)
-		}
-		if statusCode != http.StatusRequestEntityTooLarge {
-			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, statusCode)
-		}
-	})
-
-	t.Run("throw a bomb with no content length header", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: generateRandomString(200)}
-		statusCode, err := createRequestBombWithCustomHeader(ts.URL, client, adminToken, createCertificateRequestRequest, "")
-		if err != nil {
-			t.Fatalf("couldn't get status: %s", err)
-		}
-		if statusCode != http.StatusRequestEntityTooLarge {
-			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, statusCode)
-		}
-	})
-
-	t.Run("throw a bomb with fake content length header", func(t *testing.T) {
-		createCertificateRequestRequest := CreateCertificateRequestParams{CSR: generateRandomString(200)}
-		statusCode, err := createRequestBombWithCustomHeader(ts.URL, client, adminToken, createCertificateRequestRequest, "2")
-		if err != nil {
-			t.Fatalf("couldn't get status: %s", err)
-		}
-		if statusCode != http.StatusRequestEntityTooLarge {
-			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, statusCode)
-		}
-	})
 }
 
 func generateRandomString(kilobytes int) string {
