+++ internal/testutils/testdata.go
@@ -1,6 +1,6 @@
 package testutils
 
-const (
+const ( // internal/db
 	RootCACSR = `-----BEGIN CERTIFICATE REQUEST-----
 MIIDIzCCAgsCAQAwcjELMAkGA1UEBhMCVFIxDjAMBgNVBAgTBUl6bWlyMRIwEAYD
 VQQHEwlOYXJsaWRlcmUxEjAQBgNVBAoTCUNhbm9uaWNhbDERMA8GA1UECxMISWRl
@@ -395,4 +395,200 @@ y6MHDTAyVHrSouCqfh9XD6RDN58d+u9kLEg0WJD55wH4E4z+ZZhEMicCWfT/rn+b
 b3dRTVslxdJ0dOApn/6zwfRMXgI7j2yRSkA7F39ekwlPhJy2bGrEDgTlDK33AwPU
 wM1PZYERQJNOGMAI5Q==
 -----END CERTIFICATE-----`
+)
+
+const ( // internal/server
+	ExampleCSR = `-----BEGIN CERTIFICATE REQUEST-----
+MIICrjCCAZYCAQAwaTELMAkGA1UEBhMCVFIxDjAMBgNVBAgMBUl6bWlyMRIwEAYD
+VQQHDAlOYXJsaWRlcmUxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0
+ZDETMBEGA1UEAwwKYmFuYW5hLmNvbTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCC
+AQoCggEBAK+vJMxO1GTty09/E4M/RbTCPABleCuYc/uzj72KWaIvoDaanuJ4NBWM
+2aUiepxWdMNTR6oe31gLq4agLYT309tXwCeBLQnOxvBFWONmBG1qo0fQkvT5kSoq
+AO29D7hkQ0gVwg7EF3qOd0JgbDm/yvexKpYLVvWMQAngHwZRnd5vHGk6M3P7G4oG
+mIj/CL2bF6va7GWODYHb+a7jI1nkcsrk+vapc+doVszcoJ+2ryoK6JndOSGjt9SD
+uxulWZHQO32XC0btyub63pom4QxRtRXmb1mjM37XEwXJSsQO1HOnmc6ycqUK53p0
+jF8Qbs0m8y/p2NHFGTUfiyNYA3EdkjUCAwEAAaAAMA0GCSqGSIb3DQEBCwUAA4IB
+AQA+hq8kS2Y1Y6D8qH97Mnnc6Ojm61Q5YJ4MghaTD+XXbueTCx4DfK7ujYzK3IEF
+pH1AnSeJCsQeBdjT7p6nv5GcwqWXWztNKn9zibXiASK/yYKwqvQpjSjSeqGEh+Sa
+9C9SHeaPhZrJRj0i3NkqmN8moWasF9onW6MNKBX0B+pvBB+igGPcjCIFIFGUUaky
+upMXY9IG3LlWvlt+HTfuMZV+zSOZgD9oyqkh5K9XRKNq/mnNz/1llUCBZRmfeRBY
++sJ4M6MJRztiyX4/Fjb8UHQviH931rkiEGtG826IvWIyiRSnAeE8B/VzL0GlT9Zq
+ge6lFRxB1FlDuU4Blef8FnOI
+-----END CERTIFICATE REQUEST-----
+`
+	ExampleCSRCertificate = `-----BEGIN CERTIFICATE-----
+MIIEUTCCAjkCFE8lmuBE85/RPw2M17Kzl93O+9IIMA0GCSqGSIb3DQEBCwUAMGEx
+CzAJBgNVBAYTAlRSMQ4wDAYDVQQIDAVJem1pcjESMBAGA1UEBwwJTmFybGlkZXJl
+MSEwHwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQxCzAJBgNVBAMMAm1l
+MB4XDTI0MDYyODA4NDIyMFoXDTI1MDYyODA4NDIyMFowaTELMAkGA1UEBhMCVFIx
+DjAMBgNVBAgMBUl6bWlyMRIwEAYDVQQHDAlOYXJsaWRlcmUxITAfBgNVBAoMGElu
+dGVybmV0IFdpZGdpdHMgUHR5IEx0ZDETMBEGA1UEAwwKYmFuYW5hLmNvbTCCASIw
+DQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAK+vJMxO1GTty09/E4M/RbTCPABl
+eCuYc/uzj72KWaIvoDaanuJ4NBWM2aUiepxWdMNTR6oe31gLq4agLYT309tXwCeB
+LQnOxvBFWONmBG1qo0fQkvT5kSoqAO29D7hkQ0gVwg7EF3qOd0JgbDm/yvexKpYL
+VvWMQAngHwZRnd5vHGk6M3P7G4oGmIj/CL2bF6va7GWODYHb+a7jI1nkcsrk+vap
+c+doVszcoJ+2ryoK6JndOSGjt9SDuxulWZHQO32XC0btyub63pom4QxRtRXmb1mj
+M37XEwXJSsQO1HOnmc6ycqUK53p0jF8Qbs0m8y/p2NHFGTUfiyNYA3EdkjUCAwEA
+ATANBgkqhkiG9w0BAQsFAAOCAgEAVZJZD0/ojZSOVIesZvrjLG0agSp0tsXY+hEt
+I/knpYLvRcAd8b3Jx9gk+ug+FwDQ4IBIkTX18qhK2fgVUuMR/ubfpQeCMbp64N3Q
+kmN/E1eu0bl6hhHAL7jEbi0DE3vAN9huQxAIu5pCyLvZIrPJtvuyj2jOpJBZwGoP
+539lfEM++XALzI4qKQ6Z0a0rJZ4HoruKiYwEFZ7VkmRLD0uef6NMZRqa/Vx+o0uT
+1TjH4AeDDmJmP/aHlHbpXkHQ9h9rfTa6Qbypo+T9pGDhd02O1tEqrHfiQyNWJxb0
+rbR+owT32iCfayzKKqhmAYSF2d9XKWEhulgxWDaXgvUbq4Y+fgfU2qMVz5uusTDh
+a9Mp9dsYWySWEUcEa4v2w6FfaaVXE1S9ubm+HoIVtotuutL5fn86q19pAAePYjLQ
+ybiETp5LU3chuYmMlCiDRNGHYhN5nvGcttqRdWIBe454RRPNo4iGVl13l6aG8rmI
+xDfk5lIwObalbELv+mEIGI1j/j4//nJFXByxlLHm5/BF8rmvHDj1aPtPRw9DLgSX
+ejhjjec1xnkBR+JF0g474hLdPjCnA0aqLQInZbjJJm5iXzyXBg1cy7KvIBy3ZkrR
+Pp7ObjaWxjCT3O6nEH3w6Ozsyg2cHXQIdVXLvNnV1bxUbPnfhQosKGKgU6s+lcLM
+SRhHB2k=
+-----END CERTIFICATE-----`
+	ExampleCSRIssuerCertificate = `-----BEGIN CERTIFICATE-----
+MIIFozCCA4ugAwIBAgIUDjtO3bEluUX3tzvrckATlycRVfwwDQYJKoZIhvcNAQEL
+BQAwYTELMAkGA1UEBhMCVFIxDjAMBgNVBAgMBUl6bWlyMRIwEAYDVQQHDAlOYXJs
+aWRlcmUxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDELMAkGA1UE
+AwwCbWUwHhcNMjQwNjI4MDYwNTQ5WhcNMzQwNjI2MDYwNTQ5WjBhMQswCQYDVQQG
+EwJUUjEOMAwGA1UECAwFSXptaXIxEjAQBgNVBAcMCU5hcmxpZGVyZTEhMB8GA1UE
+CgwYSW50ZXJuZXQgV2lkZ2l0cyBQdHkgTHRkMQswCQYDVQQDDAJtZTCCAiIwDQYJ
+KoZIhvcNAQEBBQADggIPADCCAgoCggIBAJU+5YaFlpn+bWvVri5L6EkmbAPuavsI
+/KXY7ufRmc5qb08o1na9lLJ/7TuMD4K36Idnq20n1JohSlrdymBpNZ8O3m5fYYtk
+hx5WADlBZsKnC5aZJIChEb4bYcOFLP+d3PooVsAKBxW0Q6TECviQcK7GxaxEZw0L
+7FRhX2c9+CxbvRGP6OGVggXZxwkZik/JJ9aym+fltt9QvlxQVBq/GlFYZYC+H8jV
+Z6RnUjugnWcTm9PAsQ6+EHEevAW+dWaDP+gr9AgKKz1EXbc1mVKAVOLHjb+Ue7RC
+vFoar/YxYIszD58dOSB/GuAxn+JAjWbnOu7jeX3XeWlKOagUJF9L9TgMIUWdiuJG
+8Uu/kK2MjyRFdT8opnPFAXrK7vSuMBzhRtswAlWc8xoZWeSQF+NpjU+swbg8ySYT
+LfZxVB+s/ftxnGU3RM/RWdbZhb0DAuIBsFAGCbnj+Q61/cK4i58JVjUqzLk+XOwR
+55LAyS0Y5pj9jDc5mqvS0z7ot7s2OBM1+o8e3KJgdMSXorYkv3toHMGEIUmPQZCX
+JtRCjFNgnoWeLDc+oLiN6BlPx7bS4MDN9tMPCJwF6vnxFzLAzdRqY3D7uRS3chsx
+7ClMR9MDsSxplC7tptXgv8UTzh1XZjWGCeZq0Gbe927Hmwy2q8k/BFwnR4PIVSiE
+7YAZPb0CPmrfAgMBAAGjUzBRMB0GA1UdDgQWBBRgLXukRHTovOG6g9Z5eCaeh6Sx
+aTAfBgNVHSMEGDAWgBRgLXukRHTovOG6g9Z5eCaeh6SxaTAPBgNVHRMBAf8EBTAD
+AQH/MA0GCSqGSIb3DQEBCwUAA4ICAQA9TpgTrGmnyxKB2ne76LNQadiijVPpS6/U
+OPFAX4EPJ0V5DhDreJjsZJC6Is2Q9+qsPpn/nlW7bvZUVHGodUKcE+TQWFiMtLvu
+8ifzk8x1R46aqhTyxb7WBBFfvbvdmlEENKTmTS6A/C3nYgmkfk5N7x84iTowmsVl
+Yzz9iRzxkqQ+mU3L2/Sp5nXPYWfzV9WXIJdxWcot7f4CJ79eVFu4D9hYfzcPQ9P9
+0qCBRbH/01D2E/3uTHhZPPmK2Tp1ao5SuGLppjMPX8VWVL5CMTXOj+1LF0nJJc/J
+9MrqXwtlLyKGP6HX8qALbaXwcv7db6bF+aEsgWmIEB+0ecGk9IR3XQn7I379CO3v
+J2oUCZ++lV9e2tcRehUprE1v8i+DFhPtS1iNjrO7KnDYkXimR5zI+3sGFI9/9wY0
+4PAV/roZFiEJHe5kA49vwIihJaDgy/SPIYgG/vhdj+WeIbi1ilEi12ou7VF0tyiE
+j3eXaMAL8EAKxCUZbXcuwmK9qistAYXBFFEK9M08FwLH8HM4LoPjshMg3II9Ncs8
+p3to8U99/ZeFbJRzEUF9poZ7VwxBEcgfWD1RV0+gNLC3Au2yuc4C3anknOv7Db/r
+jdzVA8yTI8cZ/RtRohp5H/s+j2tcdfB3Zt+wfS4nLxqN/kf7qv2VSdPbXyTyz/ft
+btZkbfdL5A==
+-----END CERTIFICATE-----`
+)
+
+const (
+	SelfSignedCACertificatePK = `-----BEGIN RSA PRIVATE KEY-----
+MIIJKAIBAAKCAgEAqzwshq14z01of/TCO+m1XW9k8RLiLce2aVEK1VjuWAR3p8H4
+Oh16TkFUhXnXyY+6gWs+DKKaiX5Jekz1APKGS0gGOMbGWQfjdSP16JHvA80zi6YM
+xV8vc8piCdHzK22F93tTsVzJBL1ApQK+TmhS2WjPKM9tObKafo6uNgHd78eEHVn+
+QBkwvD8HpCnA73LabRyylJ1NC8p+fHT8+t0HjNVZNsEZOzWFZJ755oSZdV949KuY
+VnZDf+Zqmmu5SgwI9ll90hAK2Imk7hEm6t+3ViRdE8MbmLRFBykLhgWZYnJ2wzG9
+/nn4HymJmSt8BzPi+RPfbXOCBvdCRe9MVDtCwHhGukOM3yF/sfQe65EBATBE2PsR
+JWGIoeaaLjK5rT6CvQnc7YKXav8GOQoAn+o8qKuvsf6XItlRALWfwwWmdEfbIX3q
+el+P2hNF4iUoytzQKIBiMUt9IAPwW5Zh7jVK7hm1Sv94P6UIkqbdMgoa/wl9U+vF
+Oi4zgjGU8o3MjyXL716taC/5O8oGBztNDNnc4B4P46+DV/NyGqp5E/DTfZpvslRp
+KUjOmi8g8JaDIypl0m2/825d0+oG5nPrHCQLwXbVWB2UpwfMcF6Jth0SbIM7BG/1
+NnKN5OczIkhoN2ilsRFFNNa+ZEPI4aGzb9J5Hf4d63CTzVtB38Sdnh7rMxUCAwEA
+AQKCAgA3AofTZFtRTa7qnHjhwnzvXV/ySny69FPXlZ+DVqSLRjQigp/6G4o1Jau/
+jZsTN8dU2F0AtiQrU5TLY3m6Ki+Wc7b7+m+yHmSmNz1Cz88XS47pqBimN6QI8NV6
+DiaupurIzKfgbMxvZ1UjLbRxf/ZNvev/UlPmm0girDevRf5Ej7YTr9uMQAt2DlGo
+HXnL8vvU3clJERe1WAI2fWbgOK26Qrf8bSBr8w+9eY0Szzp1iIxVHeM8s5WPuzg4
+D69g6Gjgq3NOTrUNR3riuEPmZKSDWf9E4AM3lYIvgLIhBh92jws0PgCphgl+CVwZ
+Bu07ayHZntbCVL9K+Zgi23OMA9W3dK01f/rTZ4ExfJfN5KUskJ9PfaHK4J/OGAhs
+a0h0CUzhs19uxRPBwjrnIfi2E0EyIqwl+mhDAF3pxLhuTQuVyXYkTER3AQJ1QCGN
+3UP5t/HJKUUxN9jvWvv97XVNVSlkxlWiANKlvu7dvMxiRN9eKqPvzx7R4Oly3Cs8
+jd/X1uw69OFBQ3s+FAtIz8TzKS/ZdbJitZ/sz+L31qrnxb/+E19NQFh0FWElGr22
+pA0ililB4uEJFr4wHGcdyLbdt2vHsZXN5XFkRlFxz2yw57gcffoKz84w2MCtU/Qw
+Z3DgwAFmP8vVpqvTSbiQHSzL2sW/hSN0BCVK+yeULP0dm2d9aQKCAQEAxa2xWcPC
+wKyUI4022kIebd5COLKemmAnW95DShKZJqwyzAqLrHciG5QFvECl3QGmCtoty1nY
+RT1wou6NA6/3yiBJiADqsH/nV6KH1Row9bxt0sVe9ze44tgIxJtqMp4x9E1xWo9L
+zk4jzbufxUz17NvwR9ItiHpDOh88M7U0hBiUxWqBbX/OxCmZTJBdmoGaCk/G/iST
+q7BvNI+4geInE6uzLpfLiWPiFhTSW0D/0Hh9J3hdBDGheuhY7e49RHs5dUsxiVCs
+j6cJ04DvsWyZbiEsAbtO9/eJjGyB+kDUXcrlDDtharhbKxiiXDxEUa+P5+LaANGL
+a++ou4EDsPzI0wKCAQEA3cFBXbtap1dALeM3IMkJQgZ7K7N26tMvbJ8efeJZyu6a
+YsWjOPMjuuY0/tBwlH+Yet6ahIhxQuor0yVRqgEm5uIxl+NYOEw0wBQeCq20YumY
+pNRPb0+zAPQ5UOvoZCiQqwfJzq5czVu0YOx4pW0jbJzJnnhZ0B5p0gIeOgjjVynu
+lkq67c24B5Iy/wiFu/nlGc6/MDtOHpBUvtuWbADIRXptBctICI0+0sHVNMSfeyXf
+czuyPNoWb7Ze1YaB8voPRdT14ObLLGwFCjzirGtu9mFbN4WZemxCvJcm/DWimZob
+WyxXJhDbk++femuCRsS2THkpzvlM/omlEc1RZRcjdwKCAQBDFdQnM4FHZAoOGqFv
+5ppvDMuKdEvQ9irFSaOqYq9o6W1/w2BtUizYER71KTdgzmtsKWj7Ju13agdss+pV
+QwWjqdtqdW0wIuf+3KCeWHofGyhmLCczXMy45znqhxe+P+OSFioO4qyGQgxyiGcL
+TTf0fxuHNDPRqjRgaDNFFQzSe4kZijCMWaBw//EPg7rQcYU2VKainwUicgj7XH7w
+TTCXw2BWwpsHcEdM2Roeb+ug8xL+LyHaB4HWtT1g2cYfFHaGcNNJ14AIbFawWYR+
+wO0867MEj22YR4B9kGF37UJk5jNfOFOyJQiDkqOfC7Dfy+XZeyoitvpK0hWANKNt
+EAyTAoIBAQDN3l360LaGaF/yueAyFbD8lNvAZGnf85Mxej9qirrlxMGbNPQlKMRg
+/NryPTxnAFXkq8gzhh/wCUoKSbkY5Nzit9dmtO7vxP+r2oFRVJYExeyqCVh9dDYw
+ioqzb29dnLNBtIdL01/gXmSFHsOagEnEyYH8Fqr6pWGET+cT5bB4+TrAWDxWDQfu
+L3IchiLxsjtYzDF+a3BIu6GRVna9abSDm/aezGvhVI+gRcbTqD6Oq0hYyHDeQXFW
+4K4F+Uum8TNAl0Z1No4kfVvod0HQ5CQto5B1aRhlKSCDyKeuuXRsuT7wU6fMdRYE
+rw8VLb3SSUkckdeYiKVMISkX787C1MpdAoIBAHw901UFSyy3D78Z0KorBCOi0+ld
+tWr78LXI3aqJk1n3MsybT/9lETx6sE601PW+ObbsSs5y0L1S8vGOH2C4Yjk7Xqdi
+R0kpblMpY6Nh8wyeRn7F9jB4eDOLLYH6793YbjrMzJ14/1fZCfHgy+/xm/QftIP1
+W2OK+pXnlROQiyUPakO5h18x+H1S3+cnQaQXw3S5yRExC6WbmjTC8/q1yB5kEEx7
+cYYP3HUaPT4dKjFM9qt9aF//n8NJGDRw1AVd6MHOPliusM3MDHhRrWsMq2P4yXpM
+9gyV6zS00uKIH4fD5HwldtZjeLGomJK7bkFSMRmMH2jUQvnjtE4pHxJJsUw=
+-----END RSA PRIVATE KEY-----`
+	SelfSignedCACertificate = `-----BEGIN CERTIFICATE-----
+MIIGHzCCBAegAwIBAgIIGCEGnc3aBTkwDQYJKoZIhvcNAQELBQAwgaIxGDAWBgNV
+BAYTD1Rlc3RpbmcgQ291bnRyeTEWMBQGA1UECBMNVGVzdGluZyBTdGF0ZTEZMBcG
+A1UEBxMQVGVzdGluZyBMb2NhbGl0eTEdMBsGA1UEChMUVGVzdGluZyBPcmdhbml6
+YXRpb24xEzARBgNVBAsTClRlc3RpbmcgT1UxHzAdBgNVBAMTFlRlc3RpbmcgU2Vs
+ZiBTaWduZWQgQ0EwHhcNMjUwMjA0MTQxMDA1WhcNMzUwMjA0MTQxMDA1WjCBojEY
+MBYGA1UEBhMPVGVzdGluZyBDb3VudHJ5MRYwFAYDVQQIEw1UZXN0aW5nIFN0YXRl
+MRkwFwYDVQQHExBUZXN0aW5nIExvY2FsaXR5MR0wGwYDVQQKExRUZXN0aW5nIE9y
+Z2FuaXphdGlvbjETMBEGA1UECxMKVGVzdGluZyBPVTEfMB0GA1UEAxMWVGVzdGlu
+ZyBTZWxmIFNpZ25lZCBDQTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIB
+AKs8LIateM9NaH/0wjvptV1vZPES4i3HtmlRCtVY7lgEd6fB+Dodek5BVIV518mP
+uoFrPgyimol+SXpM9QDyhktIBjjGxlkH43Uj9eiR7wPNM4umDMVfL3PKYgnR8ytt
+hfd7U7FcyQS9QKUCvk5oUtlozyjPbTmymn6OrjYB3e/HhB1Z/kAZMLw/B6QpwO9y
+2m0cspSdTQvKfnx0/PrdB4zVWTbBGTs1hWSe+eaEmXVfePSrmFZ2Q3/mappruUoM
+CPZZfdIQCtiJpO4RJurft1YkXRPDG5i0RQcpC4YFmWJydsMxvf55+B8piZkrfAcz
+4vkT321zggb3QkXvTFQ7QsB4RrpDjN8hf7H0HuuRAQEwRNj7ESVhiKHmmi4yua0+
+gr0J3O2Cl2r/BjkKAJ/qPKirr7H+lyLZUQC1n8MFpnRH2yF96npfj9oTReIlKMrc
+0CiAYjFLfSAD8FuWYe41Su4ZtUr/eD+lCJKm3TIKGv8JfVPrxTouM4IxlPKNzI8l
+y+9erWgv+TvKBgc7TQzZ3OAeD+Ovg1fzchqqeRPw032ab7JUaSlIzpovIPCWgyMq
+ZdJtv/NuXdPqBuZz6xwkC8F21VgdlKcHzHBeibYdEmyDOwRv9TZyjeTnMyJIaDdo
+pbERRTTWvmRDyOGhs2/SeR3+Hetwk81bQd/EnZ4e6zMVAgMBAAGjVzBVMA4GA1Ud
+DwEB/wQEAwIBBjATBgNVHSUEDDAKBggrBgEFBQcDATAPBgNVHRMBAf8EBTADAQH/
+MB0GA1UdDgQWBBSbXLqZuUAgHvDfdpdDNDggifOgETANBgkqhkiG9w0BAQsFAAOC
+AgEAKlIcm2sMw6a9PoNFUXLHkGo46Lao1uRJ51tNGBXasbkznb3N8pr0vhVWQk7W
+QM8nd8t8yvaQ7ELo/SapBkjTyu8SwPRVbnXjl1Wke3A34VMusjhIGPNsjgI6zq8i
+rmzzFSNGb+T0AngOc6dMC9N2pAWFHnmzty7Oi39R/jhOmD38wWCMyS1Lui5tXIMJ
+hk/PnVEMF3Rx4rV8FBZh3IJE5O2hB4OCTQVh3w2kOI5+YRWTJm223WfHz7XHR2Ny
+E6yjQhPeglxspeNufhD1H8B49/XuIsHqOBrIPUxztyuBJwXz0KuFitbLrRHLaz2T
+9Ox0i6jEB5f4RU1AiwpM6KE7bFkXmWIYeOykQq7Rj0qGdo50MABGAErIIXpcRzvn
+0r3y1PCvjntdrc9IG8b0bQhVss5Irr9+5wdn90lnKAi2wASmRxbYNVuIi6BpXJvk
+lFukciXd3eW7qZrgUFzS2nBwXv7wlT6e7pfGChSrE+dlh6QBHkDlHU2FdwdAc8BX
+GbaLCYgW2R7/cgFG25APKkLxVzuIVYCQzIc94vyTqdhEhdQfvDTnnE/Nj//lVJBp
+vpdEAlIxCc8EPphBf2okYk9nbda/3fTIBlYhXo3jGw/luYY+rv3YBMzPJ9KXkfxH
+nCSo5Rj3yTrtNYFLnu+iLCvMb5PcJXE55Pu5OYGktHnvMgc=
+-----END CERTIFICATE-----`
+)
+
+const (
+	TestServerCertificate = `
+-----BEGIN CERTIFICATE-----
+MIICuzCCAaOgAwIBAgIIaw0nhrqWhFUwDQYJKoZIhvcNAQELBQAwcjELMAkGA1UE
+BhMCVFIxDjAMBgNVBAgTBUl6bWlyMRIwEAYDVQQHEwlOYXJsaWRlcmUxEjAQBgNV
+BAoTCUNhbm9uaWNhbDERMA8GA1UECxMISWRlbnRpdHkxGDAWBgNVBAMTD1Rlc3Rp
+bmcgUm9vdCBDQTAeFw0yNTA3MDEyMDQ4MDBaFw0zNTA3MDEyMDQ4MDBaMDExCzAJ
+BgNVBAYTAlRSMQ4wDAYDVQQIEwVJem1pcjESMBAGA1UEChMJQ2Fub25pY2FsMCow
+BQYDK2VwAyEAcCrx0kYogfeauM08vjntGDcmwICy3jaYqc/S2diQ9NyjgY8wgYww
+HwYDVR0jBBgwFoAUBd3se3z+TC8S6pvnWM3T7T9Ha6AwDAYDVR0TAQH/BAIwADAd
+BgNVHQ4EFgQUfQRtg1/FrrZcvhM+hLJ54nV1Q9wwCwYDVR0PBAQDAgPoMBMGA1Ud
+JQQMMAoGCCsGAQUFBwMBMBoGA1UdEQQTMBGCCWxvY2FsaG9zdIcEfwAAATANBgkq
+hkiG9w0BAQsFAAOCAQEAgfIygbfAht+GRsNFgcdowDaHGvCRCqt0mgpa4e9QMhnJ
+JJKwxhoVSVK06T2DdtCIoyexI7A3vqjLl4HGXRkrGriPKaND/Rrnrmg8DTXM2RjK
+aExl0PDFvfdnb10H6nvxeIvy03B3pY6Ch4/kxScqHJjP8FQaTSqakvY0aU204MnY
+ZUFn0EF15p9OHqXfZvJMo9njD0r/ntp7vLObsN2AdCOu0a/OAm9SGykqB/M/7j7p
+p0zicdpZMIWMv2rnP2KsLViskxelNhGtKYhQkNMvkFRxScfZPY7M0ATphuuhU3It
+BsIp/dqoLhEussA3c3dv9sfRQJAvWJXIsR2+1vA92A==
+-----END CERTIFICATE-----
+`
+	TestServerKey = `
+-----BEGIN PRIVATE KEY-----
+MC4CAQAwBQYDK2VwBCIEIJRf1zYhpb0ICETgTdil2Ad58FPiOdPDRKwQId+6Tr3Z
+-----END PRIVATE KEY-----
+`
 )
